Support for a lot more tables, filter by column, refactor/rewrite, in progress work for TBI files

+++ GT-SpecDB-Editor/Mapping/Tables/PaintColorInfo.cs
@@ -16,12 +16,12 @@ namespace GT_SpecDB_Editor.Mapping.Tables
         {
             Columns.Add(new ColumnMetadata("LabelEng", DBColumnType.String, "UnistrDB.sdb"));
             Columns.Add(new ColumnMetadata("LabelJpn", DBColumnType.String, "UnistrDB.sdb"));
-            Columns.Add(new ColumnMetadata("ColorChip", DBColumnType.Int));
-            Columns.Add(new ColumnMetadata("ColorChip2", DBColumnType.Int));
+            Columns.Add(new ColumnMetadata("ColorChip", DBColumnType.UInt));
+            Columns.Add(new ColumnMetadata("ColorChip2", DBColumnType.UInt));
             Columns.Add(new ColumnMetadata("CCBinID", DBColumnType.Short));
             Columns.Add(new ColumnMetadata("Unk1", DBColumnType.Byte));
-            Columns.Add(new ColumnMetadata("Type", DBColumnType.Short));
-            Columns.Add(new ColumnMetadata("TunerID", DBColumnType.Short));
+            Columns.Add(new ColumnMetadata("Type", DBColumnType.Byte));
+            Columns.Add(new ColumnMetadata("TunerID", DBColumnType.Byte));
         }
     }
 }
